release 3.1.184 source code

+++ services/bssintl/src/main/java/com/huaweicloud/sdk/bssintl/v2/model/ModSubCustomerBudgetReq.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonValue;
 
+import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -23,7 +24,7 @@ public class ModSubCustomerBudgetReq {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "budget_amount")
 
-    private Double budgetAmount;
+    private BigDecimal budgetAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "cancel_partner_frozen")
@@ -132,7 +133,7 @@ public class ModSubCustomerBudgetReq {
         this.customerId = customerId;
     }
 
-    public ModSubCustomerBudgetReq withBudgetAmount(Double budgetAmount) {
+    public ModSubCustomerBudgetReq withBudgetAmount(BigDecimal budgetAmount) {
         this.budgetAmount = budgetAmount;
         return this;
     }
@@ -143,11 +144,11 @@ public class ModSubCustomerBudgetReq {
      * maximum: 2147483647
      * @return budgetAmount
      */
-    public Double getBudgetAmount() {
+    public BigDecimal getBudgetAmount() {
         return budgetAmount;
     }
 
-    public void setBudgetAmount(Double budgetAmount) {
+    public void setBudgetAmount(BigDecimal budgetAmount) {
         this.budgetAmount = budgetAmount;
     }
